pyproject.toml
[project]
name = "aiwork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "asyncio",
  "ipywidgets",
  "numpy",
  "requests",
  "pymilvus",
  "openai",
  "openai-agents",
  "pygame",
  "duckduckgo-search",
  "nest-asyncio",
  "pydantic",
  "httpx",
  "jupyterlab",
  "ipykernel",
  "python-dotenv",
  "pip",
  "setuptools",
  "scikit-learn",
  "matplotlib",
  "wheel",
  "seaborn",
  "docling",
  "sentence-transformers",
  "tqdm",
  "pandas",
  "transformers",
  "rich"
]
